requirements/base.txt
Django==1.9.7
django-annoying==0.8.1
django-appconf==1.0.2
django-compressor==1.5
MySQL-python==1.2.5
rcssmin==1.0.6
rjsmin==1.0.12
Unipath==1.1

